crud

+++ backend/src/controllers/user/user.services.ts
@@ -5,6 +5,7 @@ import * as jwt from 'jsonwebtoken'
 import { Model } from 'mongoose';
 import { User } from './user.model';
 import { hash } from 'bcrypt'
+import { UpdateUserDto } from './update-user.dto';
 
 @Injectable()
 export class UserServices {
@@ -13,27 +14,35 @@ export class UserServices {
   ) {
   }
 
+  async allusers() {
+    const allUsers = await this.userModel.find({})
+    const result = allUsers.map((user) => {
+      const { name, id, apiKey, adm } = user
+      const formatUser = { name, id, apiKey, adm }
+
+      return formatUser
+    })
+    return result
+  }
+
   async create(doc: User) {
     try {
-      const existingUser = await this.checkUserWithEmailOrName(doc.email, doc.name);
+      const existingUser = await this.checkUserWithName(doc.name);
 
       if (existingUser) {
-        const error = new Error('Email or userName exists');
+        const error = new Error('userName exists');
         error.name = 'ConflictError';
         throw error;
       }
 
-      //
       const formattedUser = { ...doc };
       const salt = 10;
       formattedUser.password = await hash(formattedUser.password, salt);
 
       const result = await new this.userModel(formattedUser).save();
-      const { name, email, adm } = result
-      const token = jwt.sign({ name, email, adm }, process.env.SECRET)
-      return {token}
+      return result._id
     } catch (err) {
-      //
+
       if (err.name === 'ConflictError') {
         throw new HttpException(err.message, HttpStatus.CONFLICT);
       } else {
@@ -43,44 +52,49 @@ export class UserServices {
 
   }
 
-
-
-
-  async update(id: string, updateInfo: User) {
+  async update(id: string, updateUser: UpdateUserDto) {
     try {
-      const result = await this.userModel.updateOne({ _id: id }, updateInfo)
-      return result
+      const exist = await this.userModel.findOne({ name: updateUser.name })
+      if (exist) {
+        throw new HttpException('user name exists', 409)
+      }
+      await this.userModel.updateOne({ _id: id }, updateUser)
 
+      const { _id, name, adm, apiKey, } = await this.userModel.findById(id)
+      const token = jwt.sign({ _id, name, adm }, process.env.SECRET)
+      return token
     } catch (err) {
-      throw new HttpException(err.message, HttpStatus.INTERNAL_SERVER_ERROR);
+      throw new HttpException(err.message, err.status || HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
-
   async delete(id: string) {
     try {
-      const result = await this.userModel.deleteOne({ _id: id })
-      return result
+      const exist = await this.userModel.findById(id)
+      if (!exist) {
+        throw new HttpException('user not found', 404)
+      }
+      await this.userModel.deleteOne({ _id: id })
+      return
     } catch (err) {
-      throw new HttpException(err.message, HttpStatus.INTERNAL_SERVER_ERROR);
+      throw new HttpException(err.message, err.status || HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
 
-  async getUserByEmail(email: string) {
+  async getUserByName(name: string) {
 
-    const exist = await this.userModel.findOne({ email })
+    const exist = await this.userModel.findOne({ name })
     if (!exist) {
       return
     }
 
     return exist
   }
-  async checkUserWithEmailOrName(email: string, name: string) {
+  async checkUserWithName(name: string) {
     const exist = await this.userModel.findOne({
       $or: [
-        { email },
         { name }]
     })
-   
+
     return exist ? true : false
 
   }
